Refactor tests to allow for setup and testing with multiple ORMs
