+++ .gitignore
@@ -23,3 +23,4 @@ test_db
 test_db-journal
 .idea
 *.iml
+test/support/database/dump.sql
