+++ jsonapi-resources.gemspec
@@ -26,8 +26,10 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'simplecov'
   spec.add_development_dependency 'pry'
   spec.add_development_dependency 'concurrent-ruby-ext'
-  spec.add_development_dependency 'database_cleaner'
-  spec.add_dependency 'activerecord', '>= 4.1'
+  spec.add_development_dependency 'sequel'
+  spec.add_development_dependency 'sequel-rails'
+  spec.add_development_dependency 'activerecord', '>= 4.1'
+  spec.add_dependency 'activesupport', '>= 4.1'
   spec.add_dependency 'railties', '>= 4.1'
   spec.add_dependency 'concurrent-ruby'
 end
