+++ lib/jsonapi/resource.rb
