+++ test/support/orm/sequel/initialize.rb
@@ -1,0 +1,2 @@
+require 'sequel_rails'
+require 'sequel_rails/sequel/database/active_support_notification'
